requirements.txt
beautifulsoup4==4.11.1
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cssselect==1.1.0
emoji==0.6.0
emot==2.1
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.8.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
jieba3k==0.35.1
joblib==1.1.0
lxml==4.9.1
newspaper3k==0.2.8
nltk==3.7
numpy==1.23.2
pandas==1.4.3
Pillow==9.2.0
pygal==3.0.0
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
regex==2022.8.17
requests==2.28.1
requests-file==1.5.1
rfc3986==1.5.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
Text2Emotion==0.0.1
textblob==0.17.1
tinysegmenter==0.3
tldextract==3.3.1
tqdm==4.64.0
translatte==0.1
urllib3==1.26.12
vader-multi==3.2.2.1
